+++ xmake/modules/core/tools/cosmocxx.lua
@@ -15,10 +15,10 @@
 -- Copyright (C) 2015-present, TBOOX Open Source Group.
 --
 -- @author      ruki
--- @file        clangxx.lua
+-- @file        cosmocxx.lua
 --
 
--- inherit clang
-inherit("clang")
+-- inherit cosmocc
+inherit("cosmocc")